refactor: LocationGroupHeader.tsxを単一責任原則に従って分割
LocationGroupHeader.tsx (733行) を以下のように分割してコードの保守性を向上:

新しいコンポーネント:
- ShareDialog.tsx - シェア機能ダイアログ (214行)
- PlatformBadge.tsx - プラットフォーム表示バッジ (27行)
- PlayerList.tsx - プレイヤーリスト表示 (51行)

新しいカスタムフック:
- useShareActions.ts - シェア機能の状態管理 (35行)
- usePlayerListDisplay.ts - プレイヤーリスト表示ロジック (122行)

テストファイル:
- PlatformBadge.test.tsx - 5つのテストケース
- PlayerList.test.tsx - 5つのテストケース

メインコンポーネント:
- index.tsx - 733行から380行に削減、単一責任に集中

その他:
- docs/refactoring-plan.md - リファクタリング計画書を追加
- .devcontainer/devcontainer.json - ripgrepをインストール設定に追加

改善効果:
- 保守性: 機能ごとに責任が分離
- テスト容易性: 各コンポーネントが独立してテスト可能
- 再利用性: PlatformBadge、PlayerListは他の場所でも使用可能
- 可読性: ファイルサイズ削減により理解しやすくなった
